added the student view
