+++ edx_reverification_block/xblock/static/xml/reverification_block_example.xml
@@ -1,0 +1,1 @@
+<edx-reverification-block />
