--- reverification/xblock/static/xml/reverification_block_example.xml
@@ -1,1 +1,0 @@
-<reverification />
